+++ noodles-cram/src/writer/options.rs
@@ -2,7 +2,6 @@
 pub struct Options {
     pub preserve_read_names: bool,
     pub encode_alignment_start_positions_as_deltas: bool,
-    pub require_reference_sequence: bool,
 }
 
 impl Default for Options {
@@ -10,7 +9,6 @@ impl Default for Options {
         Self {
             preserve_read_names: true,
             encode_alignment_start_positions_as_deltas: true,
-            require_reference_sequence: true,
         }
     }
 }